requirements.txt
Django==3.0.5
djangorestframework==3.11.0
matplotlib==3.2.1
numpy==1.18.2
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
pandas==1.0.3
Pillow==7.0.0
torch==1.4.0
torchvision==0.5.0
scikit-image==0.17.2
scikit-learn==0.22.2
dlib==19.6.0
imutils==0.5.3